requirements.txt
Flask==3.0.2
Werkzeug==3.0.3
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.137.0
requests==2.32.3
gunicorn==22.0.0
